callbacks delete plane: ensure directory exists

+++ src/gazeMapper/GUI/_impl/callbacks.py
@@ -152,7 +152,8 @@ def delete_plane(study_config: config.Study, plane: plane.Definition):
     # remove config directory for the plane
     path = config.guess_config_dir(study_config.working_directory)
     p_dir = path / plane.name
-    shutil.rmtree(p_dir)
+    if p_dir.is_dir():
+        shutil.rmtree(p_dir)
     # remove from known planes
     study_config.planes = [p for p in study_config.planes if p.name!=plane.name]
